Modified to work with either driving scheme.
Instead of adapting the sample "advanced" joystick code or writing out
own, it turns out that we can use the built--in arcadeControl()
function for single-joystick driving or the tankControl() function for
double-joystick tank-style driving.

We couldn't decide which was better, so we used both!  Use Btn5U (left
upper trigger) to switch modes.  The robot will intentionally stop
briefly before switching.