Committing the latest changes to the ControllerTest file.
For reasons that are not clear to me, the tankControl() and arcadeControl()
functions that we relied on and _employed successfully_ this weekend could not
be found when we attempted to compile today.  The simplest solution to this,
though the most error-prone, was to copy the definition of these functions
from the examples and rewrite them to fit our needs.

Both tankControl() and arcadeControl() have been tested and shown to work.
You may need to reverse the motors in the motors-and-ports dialog if your
robot drives backward.

While they were at it, the students threw in a new "macnum" function that will
eventually be used as an equivalent function for holonomic drive control, with
the obvious advantages of allowing for strafing as well as normal turning.  I
have been assured that both a normal rectangular drive with Mecanum wheels and
an X-drive with Mecanum *or* omniwheels should be equally controllable under
this scheme.