This serves two purposes:
1) This is working code for the mentor-bot that is currently in
progress.  The bot uses a 5-wheel C-drive instead of Mecanum, but the
new code here is not yet tested.

the desire is to make Mecanum control and control of this bot feel the
same, which willrequire a bit more coding on our part.

2) This code uses the official competition template!  The actual
production code that the students write should do the same, so this
serves as an example for that.