Committing some sample arm code with a control scheme we're happy with.
Even with our new, compact design, aluminum bars, and a
torque-friendly gearing ratio, the two 393 motors on the new arm are
not strong enough to hold the arm up under load.

Our previous arm control scheme required us to hold one button up to
provide maximum power to the arm, another button to reverse the power,
and releasing those buttons turned the arm motors off.  This has
proven disadvantageous; since the arm motors need to be firing in
order to keep the arm up under load, we shouldn't require a human to
constantly hold down a button to make the robot do what it's supposed
to do anyway.

Better to have the arm buttons accelerate the arm up or down, with the
arm's current velocity helping to maintain its position automatically.
We also added velocity decay so that releasing the arm buttons cuts
power to the arm motors eventually.  The result is an arm that "feels
right", and stays up reasonably long under load to allow us to place
caps on the poles, though the motors *do* burn out eventually if kept
up for long enough!

This code has yet to be incorporated into competition-bot.c.  But
we'll transplant it soon.