Committing tested code for 2018-11-27.
What works:
* Manual ontrol of wrist rotation (buttons 5U and 6U.)
* Manual control of wrist elevation (buttons 7U and 7D).
* Manual control of claw opening and closing (buttons 8D and 8R.)

What doesn't yet work:
* Potentiometer reading (we're still trying to figure that out.)
* There's no encoding or gyro code yet.

The hardware is much improved, with a couple of high-strength gears
added in the arm, the original arm mechanism replaced with the third
iteration, a fault wrist rotation motor removed, rubber grommets added
to the gyro to reduce vibrations -- lots of improvements.

What's left is programming the robot to take advantage of them.