The Button7L autonomous task has now been confirmed to drive straight!
The four integrated encoders are all back on the Spudnik 1.

Natan and I are using an autonomous task with a feedback loop; whenever a pair of left/right motors has an encoder click differential, we speed up the lagging motor (or slow down the speedier motor, if the laggy motor is greater than max).  Then, having confirmed that both motors are driving at the same rate, we ramp both up or down to get to the desired speed.

A combination of this code and Edwin's slow-down code from yesterday should allow the robot to drive straight during autonomous and stop on a dime (or reasonably close to one.)