Determine the true value of ENCODER_CLICKS_PER_INCH through trial.
At this point, setting targetDistanceInches to 100.0 and calling
driveStraight() will cause the robot to drive about 101 inches before
stopping.  *Slightly* increasing
DRIVE_STRAIGHT_STOPPING_DISTANCE_INCHES by a value less than 1.0 is
probably all we need to make this more accurate.

The algorithm was simple: force the robot to travel an arbitrary
number of clicks (3000), and measure the inches traveled by the back
of the robot, over and over, averaging the result.