The rotation will slow down as it gets closer to the target bearing and min is defined.
As the difference between the current bearing and target bearing gets smaller, the speed will slow down by being multiplied by .9985 so that the robot can rotate and stop without spiralling out of control. .9985 was used so that we can stop the robot from rotating without stopping too quickly or before it reaches its target bearing.