Trying to get PID control working with the quadrature encoder.
The gear ratio is throwing us for a loop here; I feel that the PID
algorithm as implemented is otherwise correct.

The key to printing the LCD properly was to put the print statements
inside opcontrol(), which is the PROS equivalent of task main(), rather
than initialize() where I had it befire.

The display buttons on the cortex now switch between goal angles rather
than drive the motor directly.  The increments are 90 degrees right now.

By counting teeth, Jim determined that our turntable gear ratio is 66
teeth to 12, which is 5.5.  Somewhere, somehow, multiplying by and
dividing by this value will be the key to getting the turntable PID
working.

Finally, note that we have a kP of 1.0 right now, and the motor is
generally stopping near the target already.  If we need kI, it's
probably going to be small.