This is the first stab at the code for the student screw sorter bot, Deep Fryer.
It consists right now of a vibration motor (which tranlates rotary
motion to linear motion using a connecting rod) and an impeller to drive
the screws that have escaped the initial trough.

Two potentiometers allow the user to adjust the speed.  A bump sensor
toggles the motors when hit.

Warning: jammed screws can come through the impeller with appreciable
speed.  Be wary, and wear eye protection when activating the robot.