Committing a .gitignore that PROS can actually use.
The problem with the previous set of .gitignores was that they were too
aggressive -- they excluded files generated by the project creation command --
"prosv5 conduct new" -- in addition to excluding files generated by the
project build command -- "prosv5 make".  That _could_ work, but it requires
PROS to have a command that can take a barebones sets of files, like what's in
./screw-sorters/french-fries right now, and turn it into a working project.

PROSv5 has no such command.

We're stuck, then: in order to build, we need to version the files generated
during project creation, and that means we need to stop ignoring those files
first.