Commiting Brandon, Jacob, Eli, Katelyn, and Natan's holonomic X-drive code.
This is interesting, because it does not use the same formula as the Mecanum code that we have already tested and gotten to work in the Spudnik 1.  It was a Youtube video in the end that gave us the wherewithal to discover the difference in signs and correct our formula.

This code has been tested and it works reliably...at least with with current test drive, which is tentatively called "Spudsplorer 1."

The only _actual_ code here is the .vex file, which can be read by Vex Coding Studio.  The .vex.cpp file is just my attempt at versioning, in plaintext, the actual code that people can actually edit.