It's not the IntakeLift that we wanted.  But it's the IntakeLift we deserve.
The IntakeLift controls now call spin() when moving forward or reverse.
When the right trigger is released, we attempt to "spin to our current
position", which is supposed to at least _try_ to halt the downward
momentum of the arm.

in practice, it doesn't do this very well, and we need ruber bands to
slow the arm's descent down.

Oh, well.  If you can get spinToPosition() to work better without
calling spin(), then please do so.