Press "A" to deploy the cube stack.
This doesn't work.  Part of it does -- the tray deployment, and the tray
lifting can clearly be timed better.  But the drive does not drive, so
the robot never backs off.

I don't know why this is happening yet.  Autonomous calls the same
mecanumDrive() function and it works reliably there.