Added a framework for scheduling autonomous commands asynchronously.
Struct ScheduledOperation consists of an operation enum, a start time,
and a power.  We can take an array of these and execute the operations
in sequence...but the work they perform is non-blocking and runs in
parallel on the robot itself.

This is what we'll use to make autonomous and programming skills easier
to write and easier to read.  We could even rewrite deployCubes()
with this.