Fixed an error in START_TIME_MILLISECONDS that only occurred in the competition.
Note the line where we declare START_TIME_MILLISECONDS.  This is what
it was before:

    const double START_TIME_MILLISECONDS = Brain.timer(msec) / 1000.0;

That was blatantly incorrect -- it was calculating the start time in
*seconds.*  The division needed to be removed.

But why didn't we notice?  Because we then took that smaller value and
subtracted it from Brain.timer(msec) again in line 31.  This resulted
in an expression not dissimilar from:

    (Brain.timer(msec) - NEARLY_ZERO) / 1000.0;

Which is the same as the *absolute* match time, in seconds, when we had
originally intended to only use the *relative* start time.

-----

The actual matches at today's competition don't actually start at 0
seconds after enabling the field.  There's an extensive setup time,
and Brain.timer() is incrementing all the while.  By the time we got to
actually running auton, the timer had elapsed far more than 15 seconds,
and since our code was effectively using absolute time instead of
relative time, our 15-second limit was always surpassed!

No wonder we didn't do anything!

Mr. Zwissler noticed the one--you people owe him a Coke.