I made the sensors find if they are on a line.
details:When middle sensor sees line it is on a line or if the left and right see the line.
still need to make 3 other functions one for if we are not on a line and one if we are on the left or right of a line.
last we are unaware of what the variable is for the green tape.