Made code for Arm Lift, Natanonmous, and changed a Port
Named after me, natan, because my name is similar to the word autonomous. We created a function that allows the liftMotor to move autward or inward at
specific time and velocity. This means in autonomous we can just call either MotorLift(Inward) or MotorLift(Outward) to control the arm easily. Created an If/Then statemnet
that controls the ArmLift, pressing one button to move the arm up and the other button to move the arm down. Created a basic autonomous function for the right side of the field.