As a test, added a function that writes to a new file each time it is invoked.
Unfortunately, the time() function on VEX V5 Brains returns the same
time each time (the Unix Epoch: 1970-01-01).  So we need another way to
guarantee uniqueness of logfiles.

We settled on looking for files on disk with the given prefix -- the
problem is that this gets slower and slower the more logfiles are in the
system.  That's a limitation I can't overcome for now, not unless I have
the POSIX filesystem enumeration functions on hand (and they don't seem
to link when we include their headers.)