Hopefully implemented the efficiency display using the new technique.
Note the use of static variables to keep track of which set of 4 motors we are
showing during the current invocation.  Note also the use of a static variable
to keep track of when the function was last called.

This feels weird to me.  I don't want all of these functions to have their own
bespoke static variables!

And no, not a line of this has been tested.