Figured out why we were getting segment violations (that's what the red box was.)
The long and short of it is this: when you need to remove elements from
a list you're iterating through, use list.erase() and use it correctly
instead of assuming that list.remove() will do magical things.

I ended up having to replace the "for (auto foo: bar)" for-each loop
with an ordinary for-loop in order to get the iterator, since calling
erase() necessarily changes that iterator's value.

I did remove the LED and Brain.Screen printing code; we can put it back if needed.