Changed the Autonomous so the robot can do two tasks simultaneously.
Removed the unnecessary "parameter" member of the task struct. Changed
the break to a continue so if one parents tasks is destroyed then it
will still continue down the parent task list instead of breaking from
it. Changed the wait time to be shorter.

Changed the old Autonomous tasks so the robot will perform different
actions and will also perform tasks simultaneously. Made an assumption
about the gear ratio that will need to be checked later.