Removed the block that prevented two tasks from excuting at once in autonomous.
We had to change four functions by adding false since when we were trying get two tasks
to execute at the same time they wern't. We checked Vexcode api to if
there was something specific with the fucntions and we found that
the wait for completion argument is automatically true. Meaning that
functions will not start the tasks simiultaneously and will wait until
the task before it is completed. Adding false to the function changes
this so two tasks can be performed at once and the motor will still
reach the rotational value all at once.