Fixed the limit code for spatula and beetle arm
Prevents the spatula from deploying when the beetle arm deployed. To do this added the else that returns the spatula to false if isLiftArmDown =
true. Not sure if this is neccesary, belive this works because intialy
spatula is set true so once control gets to the second else, preventing it to be set to false, spatula just
returns true. This can be tested by commenting out 383 - 385 and testing
if it still works.