Attempted to solve logic failure in TurnTask::done().
- Added a constructor for the new class since the gyro turn task is
meant to replace the original turn task.
- The done function is specific to the GyroTurnTask since it invokes
smartdrive::isTurning() instead of drivetrain::isDone(). It's important
that it invokes isTurning because it's the alternative method of trying
to get the robot to say that the turning task is done instead of the
robot sitting there and not performing any further tasks.
- Changed the TurnTask to GyroTurnTask in autonomous().