While enabling the Mecanum drive we discovered a error with the brain
We trouble shot the Mecanum drive and when we did that we discovered
that there is a "plus 1" error, whenever we try to use a port it
activates a port that is one greater. For example if you try to activate
port x it activates port x+1. We worked around this in software.