Confirmed that all joint angles are measured correctly.
This is important because our PID is dead out of the water if the
desired angles and the current angles aren't in the same ballpark.  I
am explicitly using relative angles here: the shoulder, wrist, and elbow
are all assumed to be at 0 regardless of what the quadrature encoders
report, and by subtracting the current encoder measurement from the
initial encoder measurement and scaling carefully, we can achieve angles
that are remarkably consistent, given the poor quality of the encoders.

- The most important values to determine were the
  *_ENCODER_CLICKS_AT_VERTICAL and *_ENCODER_CLICKS_AT_HORIZONTAL
  values; they make or break the code.  They are easy to measure when
  RobotC is in debug mode; just move the limb in question to horizontal
  or vertical and record what the corresponding encoder says.  I had to
  do that a few times.

- Our next step, God willing, is PID itself, though I'm not happy with
  the whine that I'm getting from motors that are given a low power
  value.  Maybe I'll deadzone these.