Checking in initial PID constants.
The PID values are reasonable (more so for the wrist; less so for the
elbow and shoulder); that's not the issue.

The issue right now is that the arm's X and Y values are not
constrained.  Indeed, you can even adjust X or Y so much that they flip
signs, causing the arm to temporarily go rogue.

We're supposed to be the limbs of a triangle.  It's time to teach the
code to act like it.

Next commit, next commit.