PID tweaks are done, and no, the arm still doesn't work.
It contorts itself into strange shapes as X and Y are adjusted.  I
prevented X and Y from being negative, but that hasn't stopped the arm
from obtaining a funky negative shoulder angle, causing it to lurch
backward.

I think there is an ambiguity in my math, and I need to resolve that on
paper first.