Just pushing what I have; the script is not complete.
The way VSCode was formatting the Python irritated me and I got
distracted.  Need to figure out how to get YAPF to actually respect the
{split_before_first_argument: True} Python dictionary.

According to the answers to https://stackoverflow.com/questions/45042530/how-to-configure-yapf-to-use-hanging-indentation-for-long-argument-lists,
yapf has been ignoring this setting for some time now; my actual results
bear this out.

This should work.

    python -m yapf --in-place --recursive -vv . --style="{based_on_style: yapf, split_before_first_argument: False, indent_width: 4}"

It does not.