Important bugfix: Figured out how to make the elbow angle work!
The key insight here was separating what the kinematics equations expect
(i.e., 0 + 180 + 0 degrees when the arm is vertical) from what the sensors
were actually doing (i.e., 0 + 0 + 0 when encoders were reset.)

* calculate.py: The shoulder needs to be at 90 degrees when vertical,
so change the shoulderAngle output to make that look right.

* main.c: The bug was in getAngle() this whole time.

Think about what happens when you reset the encoders with:

    getDegrees(shoulder, true);
    getDegrees(elbow, true);
    getDegrees(wrist, true);

Now all joints think their angle is 0.  But we're vertical, so the angles
need to always add up to 180 degrees, not 0!  The elbow angle should
have been measured as 180, then, when the elbow encoder was 0; simply
subtracting 180 from the measured encoder value was all that was needed
to fix this.

Now the arm correctly contracts as I adjust Y and X, but the arm is
vertically oriented when it does it rather than the horizontal
orientation I wanted.  This might be a bug in shoulderAngle, but it
might not be.  I'll figure it out.