We created a new project to test a new bot called 'Ping Pong Bot'.
- We added code to spin the flywheel at 100% and intake at 60% because
the flywheel worked most optimally at 100% and the intake was spinning too
fast to take the ping pong balls from the passive magazine system.
- There is also code to spin all the wheels in the drive train at 30% to
test if all the motors were working and spinning in the same direction
and the 30% speed so the gears wouldn't break if they spun in the wrong
direction.
- We made a wrapper function called "drive" to be able to make the drive
train simpler so the 4 motors on each side can work together.