Creating a new, blank project that actually builds.
I think by instructing Matthew to move the .vscode top-level directory
"inward", we broke the compilation of the project.  The neew project,
you will note, puts a .vscode at the top-level as a sibling of the year
folders.  I seriously think that's the way things are supposed to be.

I fixed a couple of build errors:
- We were missing a #include for hardware.h so that main could call the
  (certified worthless) vexcodeInit() function.
- We were missing a couple of namespace references.