A number of ports on our Brain are dead.
This commit documents the broken ports and the few that are working
(two, in fact: ports 10 and 18) in hardwareConstitution.h.

We are about to switch to a temporary Brain from Team 6526, and while it
has a few known dead ports, it has a lot more working ports than two!
We need seven ports for complete functionality.

The existing port assignments in hardwareConstitution.h are not
relevant; only the list of broken and working ports in the comment is.