Finish putative teleop drive scheme switching code.
This consists of two parts: a struct which captures everything we care about
from the driver inputs, and a function which takes a scheme and returns that
struct.

The third part, which has yet to be done, is to use the struct rather than
using the inputs directly in main.cpp.  Sepehr will work with me in this later
today.