The claw is broken.
It wil work initially, but at some point, for
reasons that we have not yet determined, the claw
will either fail to open or fail to close, both
times because this line:

    clawMotor.spin(directionType::fwd, CLAW_SPEED_CLOSING_SUSTAINED_PCT, percentUnits::pct);

(Or the equivalent CLAW_OPEN line) does not apply
the given power to the claw motor.  This manifests
as the motor feeling like it's going "limp"; not
quite in coasting mode, but not applying the full
power that should be applied

- The shaft is engaged.
- The motor works; otherwise, we would not be able
  to do the initial opening and closing.
- The state machine is fine.
- The speeds that we have are temporarily wrong
  (all at 80%); we can change them later, but only
  after we've figured this out.

I'm not sure how to proceed.  The Team P claw is
BROKEN until this is fixed.