A Hail Mary of sorts for the trapjaw claw.
This change eliminates the CLAW_SPEED_{OPENING,CLOSING}_SUSTAINED_PCT variables
that we were previously using -- at my own behest -- to keep the Team P
robot's claw open or closed.  We ran into some pretty insidious problems when
testing that last night: the opening or closing would work a few times, then
"get hosed": the motor would cut to a fraction of its normal power--not enough
to move the claw, but more than 0%.  This would continue for a variable length
of time -- often more than a minute.

Needless to say, during a match, this is unacceptable.

So Mr. Z and I discussed the matter and came up with a new approach.  What if
moving the motor when it _detects_ that it's not moving causes a firmware
failsafe to be tripped?  Then we must not trip it.  And we can do that my
asking the motor to sustain its own position using a PID feedback loop by
calling clawMotor.setPosition().

I also cleaned up a few more dead variables and added comments explaining the
rest of the living ones.

What I have not done is tested the code, or even *compiled* it.  (My VSCode
VEX plug-in stopped working a couple of weeks ago.)  It'll be up to the
students to take it from here.