Did testing for the claw.
- After yesterday's debacle involving the claw, Uche
changed the code based on what he and Mr. Z had
talked about.

- I pulled the code, and, after much testing,
found that the problem was when the motor became
too hot.

- As a result, me and Uche talked over Discord and
tried to increase the claw bite force, but found
that it made the motor overheat.

- As of now, we don't have a definte solution to
fixing the problem, and will come up with a solution
tommorow.