Merged two team p competition folders into one.
Team P bot was made in the summer of 2025 and was
going to be what we used. BeachBot was made in the
fall and was what the name of the bot was going to
be. So we mixed them together so that they had the
name BeachBot, and the differential drive code.